ni-119 improve multipattern and add path copying

+++ src/sockets/editor/handlers.ts
@@ -6,11 +6,11 @@ import { Configuration } from "../../lib/configuration";
 const processes: Processes = {};
 
 export const editorHandlers = (_io: Server, socket: Socket, config: Configuration) => {
-    const edit = (data: { key: string; value: string; fileKey: string }) => {
+    const edit = (data: { key: string; value: string; fileKey: string; path: string }) => {
         const key = data.key;
         const value = data.value;
-        const fileKey = data.fileKey;
-        const fileData = config.files.find((f) => f.key === fileKey);
+        const filePath = data.path;
+        const fileData = config.files.find((f) => f.path === filePath);
 
         if (!fileData) return;
 
@@ -22,7 +22,6 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
                 value: value || "",
             },
             {
-                fileKey,
                 filePath: fileData.path,
                 indent: config.indent,
             },
@@ -30,7 +29,7 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
         );
     };
 
-    const create = (data: { key: string; value: string; fileKey: string }) => {
+    const create = (data: { key: string }) => {
         const key = data.key;
         console.log(`Create key "${key}"`);
         config.files.forEach((fileData) => {
@@ -40,7 +39,6 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
                     type: "create",
                 },
                 {
-                    fileKey: fileData.key,
                     filePath: fileData.path,
                     indent: config.indent,
                 },
@@ -49,22 +47,24 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
         });
     };
 
-    const remove = (data: { key: string; value: string; fileKey: string }) => {
+    const remove = (data: { key: string; staticPart: string }) => {
         const key = data.key;
+        const staticPart = data.staticPart;
         console.log(`Delete key "${key}"`);
         config.files.forEach((fileData) => {
-            makeChanges(
-                {
-                    key,
-                    type: "delete",
-                },
-                {
-                    fileKey: fileData.key,
-                    filePath: fileData.path,
-                    indent: config.indent,
-                },
-                processes,
-            );
+            if (fileData.staticPart === staticPart) {
+                makeChanges(
+                    {
+                        key,
+                        type: "delete",
+                    },
+                    {
+                        filePath: fileData.path,
+                        indent: config.indent,
+                    },
+                    processes,
+                );
+            }
         });
     };
